requirements.txt
customtkinter==5.2.2
argon2-cffi==20.1.0
pycryptodome==3.21.0
